requirements.txt
pyobjc-framework-Cocoa>=10.0
pynput>=1.7
